+++ src/main/java/org/scion/jpan/PathPolicy.java
@@ -17,6 +17,7 @@ package org.scion.jpan;
 import java.util.*;
 
 public interface PathPolicy {
+  String NO_PATH = "No path found to destination.";
   PathPolicy FIRST = new First();
   PathPolicy MAX_BANDWIDTH = new MaxBandwith();
   PathPolicy MIN_LATENCY = new MinLatency();
@@ -24,21 +25,21 @@ public interface PathPolicy {
   PathPolicy DEFAULT = MIN_HOPS;
 
   class First implements PathPolicy {
-    public RequestPath filter(List<RequestPath> paths) {
-      return paths.stream().findFirst().orElseThrow(NoSuchElementException::new);
+    public Path filter(List<Path> paths) {
+      return paths.stream().findFirst().orElseThrow(() -> new NoSuchElementException(NO_PATH));
     }
   }
 
   class MaxBandwith implements PathPolicy {
-    public RequestPath filter(List<RequestPath> paths) {
+    public Path filter(List<Path> paths) {
       return paths.stream()
           .max(Comparator.comparing(path -> Collections.min(path.getMetadata().getBandwidthList())))
-          .orElseThrow(NoSuchElementException::new);
+          .orElseThrow(() -> new NoSuchElementException(NO_PATH));
     }
   }
 
   class MinLatency implements PathPolicy {
-    public RequestPath filter(List<RequestPath> paths) {
+    public Path filter(List<Path> paths) {
       // A 0-value indicates that the AS did not announce a latency for this hop.
       // We use Integer.MAX_VALUE for comparison of these ASes.
       return paths.stream()
@@ -48,15 +49,15 @@ public interface PathPolicy {
                       path.getMetadata().getLatencyList().stream()
                           .mapToLong(l -> l > 0 ? l : Integer.MAX_VALUE)
                           .reduce(0, Long::sum)))
-          .orElseThrow(NoSuchElementException::new);
+          .orElseThrow(() -> new NoSuchElementException(NO_PATH));
     }
   }
 
   class MinHopCount implements PathPolicy {
-    public RequestPath filter(List<RequestPath> paths) {
+    public Path filter(List<Path> paths) {
       return paths.stream()
           .min(Comparator.comparing(path -> path.getMetadata().getInternalHopsList().size()))
-          .orElseThrow(NoSuchElementException::new);
+          .orElseThrow(() -> new NoSuchElementException(NO_PATH));
     }
   }
 
@@ -68,14 +69,14 @@ public interface PathPolicy {
     }
 
     @Override
-    public RequestPath filter(List<RequestPath> paths) {
+    public Path filter(List<Path> paths) {
       return paths.stream()
           .filter(this::checkPath)
           .findAny()
-          .orElseThrow(NoSuchElementException::new);
+          .orElseThrow(() -> new NoSuchElementException(NO_PATH));
     }
 
-    private boolean checkPath(RequestPath path) {
+    private boolean checkPath(Path path) {
       for (PathMetadata.PathInterface pif : path.getMetadata().getInterfacesList()) {
         int isd = (int) (pif.getIsdAs() >>> 48);
         if (!allowedIsds.contains(isd)) {
@@ -94,14 +95,14 @@ public interface PathPolicy {
     }
 
     @Override
-    public RequestPath filter(List<RequestPath> paths) {
+    public Path filter(List<Path> paths) {
       return paths.stream()
           .filter(this::checkPath)
           .findAny()
-          .orElseThrow(NoSuchElementException::new);
+          .orElseThrow(() -> new NoSuchElementException(NO_PATH));
     }
 
-    private boolean checkPath(RequestPath path) {
+    private boolean checkPath(Path path) {
       for (PathMetadata.PathInterface pif : path.getMetadata().getInterfacesList()) {
         int isd = (int) (pif.getIsdAs() >>> 48);
         if (disallowedIsds.contains(isd)) {
@@ -117,5 +118,5 @@ public interface PathPolicy {
    * @return The "best" path according to the filter's policy.
    * @throws NoSuchElementException if no matching path could be found.
    */
-  RequestPath filter(List<RequestPath> paths);
+  Path filter(List<Path> paths);
 }